Migrate can_disable_individual_chapters to BookTest
