+++ tests/testsuite/search/disable_search_chapter/src/SUMMARY.md
@@ -1,0 +1,6 @@
+# Summary
+
+- [Keep Me](first/keep_me.md)
+- [Disable Me](first/disable_me.md)
+- [Second](second.md)
+    - [Second Nested](second/nested.md)
