+++ tests/testsuite/search/disable_search_chapter/src/first/disable_me.md
@@ -1,0 +1,1 @@
+# Disable Me
