+++ tests/testsuite/search/disable_search_chapter/src/first/keep_me.md
@@ -1,0 +1,1 @@
+# Keep Me
