+++ tests/testsuite/search/disable_search_chapter/src/second.md
@@ -1,0 +1,1 @@
+# Second
