+++ tests/testsuite/search/disable_search_chapter/src/second/nested.md
@@ -1,0 +1,1 @@
+# Second Nested
